Added MoleculePattern class to pattern module.
The MoleculePattern class is actually the equivalent of the ChemGraph class,
as both derive from the graph.Graph class. However, there is no wrapper class
for it the way Molecule is a collection of ChemGraph resonance forms. There is
some duplication between MoleculePattern and ChemGraph, mainly in alias
methods for working with the atoms and bonds.