Merged ChemGraph and Molecule classes.
In writing the unit tests for molecule objects, I realized it was unnecessary
to have this extra level of splitting between these two classes. This also
improves the parallelism with the pattern module. The only expense is that the
merged Molecule class has a lot of methods, and is therefore a bit large.

Updated moleculeTest.py to match this change; all unit tests again passed.