Some bugfixes encountered while attempting to compile modules using Cython.
There is a bug in Cython which requires that certain "from <module> import
<identifier>" lines be commented out when compiling in order for the
compilation to succeed. These lines are clearly needed in pure Python mode,
but seem to not be needed for compiled Cython mode. I have mentioned this to
the Cython devs, hoping they will fix it for the pending Cython 0.13 release.

There is also a bug when using derived extension types, as Cython does not
properly resolve the derived class overriding method. I mentioned this to the
Cython devs a few months ago, but don't know if it was solved for Cython 0.13.