Refactored base module to use ChemPy library.
Most of the refactoring involves the Dictionary class, which now uses
Molecule or MoleculePattern objects instead of Structure objects. Adjacency
lists in the dictionary are now automatically converted to the appropriate
object as the dictionary is read, rather than all at the end.

Added a unit test for the base classes that uses the old-style RMG thermo
database. This unit test currently passes.