Significant refactoring of thermo module to use ChemPy base library.
The ThermoDatabase and ThermoDatabaseSet classes were merged together, with
the combined class being named ThermoDatabase but acting primarily like
ThermoDatabaseSet. Also, primaryDatabase was removed from ThermoDatabase, as
the latter is meant to chiefly generate group additivity-based values, and
therefore to be called after all primary thermo databases are checked for the
species of interest.

The symmetry number correction to entropy is coded, but the input symmetry
number cannot yet be automatically estimated. Also, conversion from Wilhoit
to NASA polynomials is not yet working.