Fix for unsymmetric Eckart tunneling correction.
To get the proper tunneling coefficient, you must be working in the exothermic
direction (i.e. the direction with the smaller barrier). The fix checks that
this is the case and, if not, swaps the barriers so that it is. (The tunneling
coefficient for the forward and reverse directions must be equal to preserve
thermodynamic equilibrium.)