Refactored kinetics module to use ChemPy base library.
The previous implementation had quite a bit of vestigial code from when I
thought we should allow the possibility of fitting rates in either direction.
Since this is not something we wish to do, most of that code has been removed.

Many other changes stem from the desire to read the original RMG Java-style
database instead of the modified version currently used in RMG Py. Eventually
we will switch to a Python-based format, but not yet.