Added DataEntry, ThermoEntry, and KineticsEntry classes for database entries.
There are a few common attributes to all entries in any RMG-style database:
index, label, short and long comments, history. These are collected in a base
class DataEntry. Derived classes ThermoEntry and KineticsEntry store extra
attributes for the thermo or kinetics model to use or a string label for a
different entry that contains the data to use.

The loaded thermo and kinetics databases now expect that each value in the
library be a ThermoEntry or KineticsEntry object, respectively. The unit tests
have been updated accordingly.