Added new-style thermo group additivity database.
Unlike the old-style RMG Java database, each part of the new-style thermo
database is contained within a single Python script, combining the dictionary,
tree, and library. The top of each file contains the tree, followed by entries
for each node that act as the dictionary and library combined. (This works
because each node has a unique entry in the library.)

The new-style database is better because it allows us flexibility in the
type of thermo data stored. For example, we can use a Wilhoit or NASA
polynomial model if such data is available. This will be more important for
the primary thermo database, something that has not yet been generated.

This should be a direct port of the old-style RMG Java database in terms of
content, i.e. the old-style database can be regenerated from the above with
no loss of information.