Changed definition of root node in other thermo database.
The prior implementation simply used a definition that would match any atom
in the molecule. This led to "dummy root" (the comment for that node) appearing
once per heavy atom, which is silly because there is no group additivity
correction for that node. By changing it to a union, we avoid these superfluous
"dummy node"s in the estimated thermo's comment.