Moved log initialization to rmg.py; refactored command-line argument parsing.
RMG Py now uses the more powerful argparse module for parsing command-line
arguments. This is a part of the Python standard library as of 2.7; users of
older versions will need to install it manually.

Also decided not to keep the custom logging module. Instead we will redefine
the levels such that logging.debug() prints with the -v flag; you must use
logging.log() to print things with only the -d flag.