Improvements made while trying to use the database in RMG Py.
Some aspects of the kinetics database had only been implemented for
MoleculePattern objects, but also needed to work for Molecule objects as well.
These have been modified to handle both cases. (Before there was no such
distinction between a Molecule and a MoleculePattern, but now there is!)

Also added a function to the rmgdata module to get its current location on
disk.