Can now generate reactions between species in input file.
The module rxngen was created for this purpose, collecting and refactoring a
number of functions from the old rmg.kinetics.data module.

The algorithm has been redesigned to generate the thermo and kinetics after
all new species and reactions have been generated, not as we go. I think this
helps compartmentalize what RMG is doing and thereby will make things easier
to follow and debug. (Thermo and kinetics generation does not yet happen,
which is why it seems to go much faster than it used to.)

Also added a summary table at the end of each enlarge() call explaining
what has been added. Again, collecting this information into hunks makes it
easier for the end user to read.