Added preliminary version of symmetry number calculation functions.
These are largely a port of the implementation in the old RMG Py. As such,
they have the same issues that those methods had. In particular, the cyclic
symmetry number calculation has not really been implemented yet, so the
symmetry number for certain cyclic species will be incorrect. It  works
enough to continue for now, but someone ought to come back to these in the
future to make them better.