Significant refactoring of the thermo_converter module.
The GA -> Wilhoit conversion was replaced with the simpler Wilhoit.fitToData()
method. Whether the latter provides a good enough fit remains to be seen; if
not, the former can be reinstated.

The Wilhoit -> NASA is largely unchanged, except that the analytical
integrals are now stored as functions in the thermo_converter module rather
than methods of the WilhoitModel and NASAPolynomial classes. This is because
they are rather unwieldy and would clutter up the chempy.thermo module.