Can now generate thermo data for species using primary and group databases.
The input file format was modified to allow for group and primary databases,
for both thermo and kinetics databases. The primary databases are loaded first
so that they take priority over the group databases.

This was needed in order to get thermodynamics of H and H2, which don't
contain any heavy atoms.