Split thermo model conversion into separate function from generateThermoData().
This is because the end user may not wish to do the conversion for every
single generated item.

Also fixed bug in thermo estimation that tramples on existing structure.
We were using a shallow copy where a deep copy was necessary. This manifested
itself in duplicate species appearing in the reaction model in RMG Py.