Can now generate kinetics for generated reactions.
Like the thermo generation, I've clumped all of the kinetics estimation to
occur subsequently, rather than immediately as the reaction is being created.
Because there is no model fitting to be done, the kinetics estimation appears
to be noticeably faster than the thermo estimation.

To enable deferred kinetics estimation, we must store the reactant molecules
with the atoms properly labeled until the kinetics are generated. Once this is
done we delete these molecules to recover the memory.