Changed methods of thermo and kinetics models to use scalars instead of numpy arrays.
Using scalars seems to be more useful, and will still give a nice speed boost
when compiled with Cython, but will run a bit slower in pure Python mode.
Also, there were a few bugs left over from when the original code was
converted from scalars to numpy arrays, so not all of the numpy array
versions worked anyway.