Commented out base class declarations of common methods in ThermoModel class.
This is to avoid a bug in derived class overriding method wherein the last
derived class implementation seems to "win" and overwrite all others. The
tradeoff is that ThermoModel objects may not be able to call these methods
without being cast, or at least requiring the full Python lookup. If Cython
addresses this bug in a future release, we will restore the base class
declarations.