Bugfixes to the thermo model fitting.
In convertGAtoWilhoit(), the method WilhoitModel.fitToData() was being called
with the parameters in the wrong order, leading to nonphysical fits
manifesting particularly in incorrect entropies and free energies. This in
turn caused Cantera simulations to fail due to absurd thermo parameters.

Also fixed a few cases where a numpy array was passed to a thermo method when
a double should be used.