Restored sortingLabel attribute to Vertex objects.
There are a number of places in the graph isomorphism code that rely upon this
value as it was previously implemented, including the sorting of pairs and
terminals. It is not terribly expensive to include in terms of either memory
or CPU time, so it makes sense to restore it.