Two minor bugfixes to the base and kinetics modules.
In Database.matchNodeToStructure(), there is a possibility that a label may
exist on multiple centers on the functional group. The code can now handle
this occurrence.

Fixed calls to isVertexInCycle() in ReactionFamily.applyRecipe() that used to
be attached to the graph attribute of the reactant structure. This is no
longer the case, so the call has been modified appropriately.