Some tweaking of the Cython declarations to provide minor speed boosts.
Molecule and MoleculePattern provide atoms and bonds as aliases for vertices
and edges, to encourage readability. These are implemented via the property
functionality, which doesn't seem to map well to Cython. Thus, using atoms
and bonds incurs the full Python penalty, and should be avoided in time
critical code.

There are many places where we wish to iterate over the dictionary-of-
dictionaries format for bonds in a Molecule (or MoleculePattern). In these
instances it may be beneficial to temporarily store the inner dictionary as
a typed Cython variable to lower the C overhead. This had already been done
in a few places, and a few more have been added.

There is a bug in Cython regarding importing of module-level cpdef methods
that were already cimported in a .pxd file. This only matters in pure Python
mode. You will need to comment out line 44 in molecule.py in order to
build that module.

Tweaked the code in atomTypesEquivalent and atomTypesSpecificCaseOf to
hopefully improve performance, or at least help identify the slow bits in
each.

Added a few Cython declarations to functions that were missing them.