Created AtomType objects to replace strings as internal representation of atom types.
There are two primary advantages to this: (1) pointer comparison instead of
string comparison and (2) caching of metadata regardin atom types to speed up
methods. In particular, the equivalent and isSpecificCaseOf methods of
Atom and AtomPattern both benefit significantly from the caching of metadata.
The memory penalty is very small since we only keep one copy of each atom type
in memory.