Updated use of callDepth in graph isomorphism.
The callDepth variable is now initialized with the number of vertices in the
smaller graph *minus those already matched in initialMap*. When callDepth is
zero, then we should have a complete isomorphism or subgraph isomorphism
match. The code has been refactored accordingly.

This lets us use the rapid test callDepth == 0 to check if we are at the
end of an isomorphism search, which provides a small speed boost.