Minor updates to the Makefile and setup.py file.
The Makefile now also removes compiled modules in the chempy.ext package. This
was needed now that we have Cythonized chempy.ext.thermo_converter.

The setup.py file contains a commented-out line useful for turning on
profiling information generation in all Cython modules. This is necessary for
pure Python mode because the usual "# cython: profile=True" does not appear
to be recognized in a .py file. It is recommended to compile with this line
commented out, then recompile only the modules you wish to profile, as it does
add quite a bit of overhead.