Index column in old-style library files now optional.
Code has been added to Library.parse() that tests for the index column in each
row and handles both cases of when it is present and when it is not present.
The test assumes that the first item is an index if it can be successfully
cast to an int without raising a TypeError.

The primaryThermoLibrary library file, which was changed in a recent commit
to have index columns, has been restored to the original form.