Use doubles instead of numpy arrays in thermo methods of Mode objects.
It has proven to be more convenient to be able to use doubles for calls to
Mode.getHeatCapacity() et al than numpy arrays. The numpy array methods have
been renamed to e.g. getHeatCapacities(), but now act as wrappers to the
double versions. The density of states methods still use numpy arrays because
of the nature of the density of states computation, requiring convolution of
each mode.