Internal molecular degrees of freedom can now be estimated.
This means that the question "What are the internal degrees of freedom for
species X?" can now be answered. The answer is an estimate, of course.

Quite a bit of work was done on the determination of the characteristic groups
and their frequencies. In particular, the determination of the groups now
more closely resembles that of the thermo database (instead of simply searching
for all subgraph matches of each node in the library). The benefit of this,
besides not having to rewrite portions of the frequency database, is no more
concern about degeneracies in the matching. Note that a few sets of children
of nodes without data are not mutually exhaustive, so it is possible to match
a group but not get any frequencies from that match.

A dependency on the statesfit module -- which implements "Frankie" -- is
required in order to fit modes that aren't determined by the characteristic
group frequencies.