Added capacity for estimating states data (internal degrees of freedom).
A recent commit in the RMG database project enables the ability to query it
for the internal degrees of freedom for individual species. This commit
enables the loading of the frequency databases and calling into the algorithm
for frequency estimation. Note that this must be done *after* the thermo
estimation since the heat capacity data is required.