Added capacity for reading pressure dependence option from input file.
This is the first step toward [re]enabling pressure dependence functionality,
which is otherwise nonfunctional.

Note that the setting unimolecularReactionNetworks has been replaced with the
more informative pressureDependence. The information stored here is the same,
however: a tuple of parameters if pressure dependence is on, or None if it is
off.