Added method for fitting reverse kinetics to Reaction objects.
This method currently only works when the forward kinetics are ArrheniusModel
objects. Note that the implementation is different than that of RMG Java's
fitReverseKineticsRoughly() in that it does not use the enthalpy of reaction
to determine the reverse activation energy. Instead, it uses a linear least-
squares fit to produce the best possible fit to a range of k(T) data points
calculated for the reverse direction. There is still a bit of error introduced
here, but probably not as much as the RMG Java implementation.