Added dummy pressure argument to ArrheniusModel.getRateCoefficient().
This allows for calling of any KineticsModel.getRateCoefficient() -- except for
ArrheniusEPModel, which is a specialty class -- using the same prototype. Of
course, the pressure is not used by ArrheniusModel, and a default argument is
provided so that the user can also call it with just the temperature with no
problems.