Added Cython header files for several modules.
Note that the Cythonization is designed to be independent, which means all
ChemPy objects must be treated as generic Python objects. This should not
serve as much of a speed penalty, as the slow bits are generally after we
have generated the needed numpy arrays.

These alone do not provide much of a speed boost in compiled mode, as we have
not declared types for any local variables yet.