Added Cython declarations to methods in msc and rs modules.
However, the speed boost is not nearly as much as I had hoped. The Cython
array access is still quite slow; we need to be able to declare the shape
and type of the arrays, which can't yet be done in pure Python mode (as of
Cython 0.13). Options are: live with the slower code, switch to .pyx files for
now, or switch to use f2py.