Improved procedure for calculating the bounding box for the molecule drawing.
Since we are using a dummy surface for the first pass, we can keep track of
the bounding box of each individual piece of the drawing, then use that
information to get the correct size for the real surface and offset the
second drawing accordingly. The adjustCoordinates() method was eliminated as
it was made obsolete by this commit. Note that the boundingRect variables
have the form [x1, y1, x2, y2], where (x1, y1) and (x2, y2) are the top left
and bottom right corners of the bounding rectangle, respectively.