Much improved procedure for drawing structure or text labels on PES drawings.
By taking advantage of temporary Cairo surfaces, we can reduce the structure
and text label drawing procedure to have nearly the same algorithm. No matter
which is used, the labels are drawn from wide to narrow as you move down the
drawing (away from the local minima itself). Multi-species configurations are
drawn with a + separating the labels or structures. A semitransparent
background is drawn behind the labels to help them stand out from the other
lines on the PES.