Now using Gauss-Chebyshev format in pressure dependence calculations for Chebyshev fitting.
This is important to getting the best possible fit to the k(T,P) grid, and
avoiding unphysical polynomial fits. As a result, I recommend using more
terms in the Chebyshev fits than before: two less than the number of
temperatures and one less than the number of pressures. This should minimize
the error contribution from the fitting while also minimizing the chance of
unphysical polynomial fit regions.