Added output file writing support.
A Pythonic-syntax output file is created in the output directory that
lists the species and net reactions and their associated parameters. A new
module output was created for this purpose. It contains methods for writing
both MEASURE input and MEASURE output files.

The output file format is still under development, although the final result
will look something like this.