Corrected return info for fminbound (returns scalar, not numpy array).
Not entirely sure what changed. I know that the original works for
Python 2.6.x running numpy 1.3.0 and scipy 0.7.0 on Ubuntu 10.04, while
the new version works for Python 2.6.x running numpy 1.4.1 and scipy
0.8.0 on Arch Linux 2010.05. Since fminbound comes from scipy.optimize,
I am tempted to believe that it is the source of the difference, but
comparing the documentation is inconclusive.