Added -o command-line flag for specifying the output file location.
This is designed as a parallel to the -d flag that is used to indicate
PES drawing mode and specify the output image file. If -o is not used,
the output file is saved to the same directory as the input file with
the name output.py.

Also added longer description to be shown when the -h flag is used.