Can now specify bath gas as a mixture of species.
The bathGas() block in the input file was moved to be a parameter of the
collisionModel() block. This parameter is a dict with the keys being the
labels and the values the mole fractions. A weighted average of the
Lennard-Jones parameters and the molecular weight will be used to
compute the collision frequency. (A weighted average could also be used
for the dEdown, but currently we make you specify that explicitly.) The
input file reading and writing has been updated accordingly.

Also fixed a few minor bugs in the input file reading and writing.