Added reStructuredText documents for remainder of Developers' Guide.
Most of these do not have much content, as they use the autodoc feature
of Sphinx to extract the docstrings directly from the Python code. This
is very nice, as it allows us to document the code and create
consistent, standalone documentation at the same time.

This commit more or less completes the first draft of the MEASURE
Developers' Guide. A bit more exposition and revision would probably be
useful.