Fixed small logic error in molecule drawing for water.
A hardcoded special case was added so that water would be drawn as H2O
and not OH2. However, there was an extra "not" in the if statement that
was causing H2O to be drawn as OH2 and CH4 to be drawn as H4C, the
exact opposite of the intended effect. This has now been fixed.