Added specialized convolve() method to reaction module.
The convolve() method in chempy.states was determined to be the
limiting step in the k(E) evaluation. Adding a version of convolve() to
reaction.pyx allows for us to declare the shape of the numpy arrays and
thereby speed up the algorithm considerably. If there is ever a way to
achieve the same speedup in pure Python mode then we can go back to
calling chempy.states.convolve().