Can now save thermo, states, and kinetics to output file.
The format of the output file is inspired by what I use in MEASURE, in that
quantities with units are presented in the form (value, units) or ([values],
units) as necessary. Each record is appended to the output file, allowing
for multiple models to be generated in a single CanTherm job.