Can now specify ground-state energy explicitly in input file.
This is useful if you use a program other than Gaussian to get a very
accurate estimate of the energy (e.g. MOLPRO). If an energy is not specified
it can still be automatically read from the Gaussian log file. Note that the
specified energy should be Hartree/particle and be on the same basis as what
Gaussian would return.