Some improvements to fromAdjacencyList().
These improvements make this method much more robust, so that it fails
gracefully (with an InvalidAdjacencyListError) if the provided adjacency list
is invalid for any of several reasons. The function now also automatically
determines whether a label is provided with the adjacency list or not, so
this is no longer a parameter.