Improved versions of __init__ and __repr__ for thermo classes.
The __init__ and __repr__ functions can now generate and process the
style to be used in the new-style database. The __init__ function can
now handle quantities provided as 2-tuples (value, units), where value
is a number or list/tuple of numbers, and units is a string. This is
done via the function chempy.constants.processQuantity(), which converts
the value(s) to SI units for internal storage. No checking that the
correct units are obtained is currently performed.

Similarly, the __repr__ method will now annotate the parameters with
their units when appropriate. This allows us to use the representation
directly in the new-style database and other flat text files.