Another bugfix in toAdjacencyList().
Generating adjacency lists for Molecule objects requires the hydrogen
atoms to be explicit. Thus we temporarily make them explicit for the
purposes of toAdjacencyList(). Before this was done within
toAdjacencyList(), which causes an exception when trying to generate an
adjacency list for a MoleculePattern. This commit allows
toAdjacencyList() to work properly in both cases.