Added portion of "long-time" method for CSE k(T,P) value determination.
Supposedly this method is a bit more numerically robust at temperatures where
one chemically-significant eigenvalue approaches the continuum of internal
energy eigenvalues. However, I have not noticed this to be the case in
practice; as far as I can tell, both the "long-time" and "initial-rate"
methods give almost exactly the same output at all conditions where you might
believe them. Thus the new long-time method is commented out.

Also note that the long-time method as currently implemented does not allow
for determining of k(T,P) values for reactions that produce product channels.
It is not clear to me how to do this without resorting to the initial rate
method.