Attempt to use DASSL for full master equation simulations.
Using scipy.integrate is equivalent to using VODE. Since we are likely
to have stiff equations, it may be that DASSL is more appropriate.
However, so far I have noticed that DASSL runs a bit slower than VODE for my
test system (Nisom = 2). Thus the DASSL code is commented out for now.