Added method for computing k(T,P) values from uij(E) and vim(E) data.
As my pressure dependence paper explains, all of the approximate methods of
reducing the master equation to a set of k(T,P) values also produce a set of
time-independent population distribution vectors uij(E) and vim(E) that can be
used along with the full master equation matrix to obtain the k(T,P) values.
This method demonstrates that this idea is in fact correct. However, it is
commented out because the individual methods can produce the same k(T,P)
values more quickly.