Changed RS method to use general procedure for extracting k(T,P) values from uij and vim.
It turns out that the general procedure, which uses all energy grains, is more
numerically robust than the specialized (and faster) procedure which only
requires some of the grains. This is particularly noticeable when the
reservoirs are small (i.e. due to a short reaction barrier). Thus we make the
switch because it is safer to do so.

To mitigate the time penalty I have moved the function
computeRateCoefficients() used for this from measure.network.Network to
measure.me and Cythonized it.