Minor tweaks to enable successful compilation with new Cython 0.14.
In particular, the new Cython release doesn't seem to like adding a Python
float (returned from a numpy function) to a numpy array. I don't think this
rewrite is much of a speed penalty due to the Cython compilation (if we were
using pure Python mode, it would be much slower, though).

Also removed some #! from the first lines of a few .pyx files, as these are
not Python files.