CSE method now uses "long-time method" to generate k(T,P) data.
The long-time method is nice because it clearly does not depend on the initial
condition. (This is also true of the previously-implemented "initial-rate"
method, but is not obvious from the implementation.) Furthermore, the
long-time method is numerically more robust as a chemical eigenvalue
approaches the continuum. The cutoff for distinctness is 2-3x rather than
5-10x. Finally, I don't think the long-time method is any more difficult to
understand or implement than the initial-rate method.

The long-time method does include an extra matrix inversion, but the matrix
is much smaller than the one we found the eigenvalues off, so the added time
should be negligible. Plus, that matrix is composed of eigenvectors, and
should always be invertible since we are dealing with physically meaningful
systems.

One piece of the initial-rate method remains, as it's the only way I currently
know to construct the uij and vim.