Added default values for CYTHON_FLAGS and CLEAN_FLAGS to Makefiles.
Curiously, up to now an end user could not obtain a working RMG Py setup
by simply checking out the repository and running the Makefile because
the compiled modules would not be put in the proper place. The values of
CYTHON_FLAGS and CLEAN_FLAGS are the ones I prefer because they keep
all build files (except the compiled modules) out of the source
directories. If these don't work for all platforms, the simpler
--inplace flag should be used instead.