Updates to the fromAdjacencyList() function.
Added a line for handling the case where the bond list is
comma-delimited instead of whitespace-delimited. The former is
considered deprecated, but must still be supported for now.

Most of the body of the function is now contained within a try block. If
an InvalidAdjacencyListError is thrown, it is caught long enough to
print the erroneous adjacency list to stdout before reraising the same
exception. This should help with debugging if this error is ever
encountered in the future.