Fixed bug regarding minimizeDensityOfStatesCalculations mode.
The bug caused a divide-by-zero error (which manifested as a warning) because
MEASURE was incorrectly assuming the reactant channels would have densities of
states, when they clearly do not in this mode. A quick check of this flag
allows us to avoid this issue. (The bug likely did not cause anything to run
incorrectly; it just gave a superfluous warning.)