Carry on if guppy package is not installed.
This effectively makes the guppy package (used in RMG for memory profiling)
into an optional dependency. I made this change because the latest version of
guppy seems to take a long time to set up the memory profiler. I'm also
wondering if there is significant overhead to its use; if this is true then we
should probably not use profiling at all (or at least allow the user to
toggle it on/off). Finally, there hasn't been a new guppy release in quite
some time, and the latest "stable" version doesn't work with Python 2.7.