Only include active K-rotor if no translations or external rotations.
The idea of the active K-rotor is to smooth out the density of states when no
external rotational data is available, e.g. in RMG calls to MEASURE. However,
this code was being inadvertently called for monatomic species, which was
causing CanTherm to give incorrect results. A workaround is to only use the
active K-rotor when there are no external rotations *and* no translational
modes available. However, we may wish to move the active K-rotor elsewhere so
that it is only used in MEASURE.