Added degeneracy attribute to Reaction objects.
This attribute is used to store the reaction path degeneracy. When the rate
coefficient for the reaction is requested, the degeneracy is automatically
multiplied into the k(T) or k(T,P) value returned from the kinetic model. Thus
it is very important that the model does not already include this correction!
This will be tricky to ensure, but fortunately most reactions have a very
small degeneracy anyway.

This replaces and supercedes the multiplier attribute used previously in RMG
only to provide the same information.