Removed keyword arguments from Graph.isIsomorphic() et al.
Using keyword arguments for a function forces Cython to call the Python
version of that function rather than the Cython equivalent (at least as of
version 0.14). In this instance, the small decrease in readability is easily
made up for in speed. My test case is instantly 20-25% faster just from this
change alone.