Now using dict of lists for global species list in RMG.
The dict of lists provides a means of grouping species so that, when checking
for whether a species already exists, we only check a small subset of the
overall set of existing species. In this application we are using the chemical
formula as the key of the outer dict because it is very rapid to generate, yet
generates a lot of "bins" to sort species into. The values corresponding to
each formula are a list of species with that formula, in order of creation.

In my test case the checkForExistingSpecies method went from 41% of the time
to 5% of the time, suggesting a speed boost of ~1.5x, which is very
significant for such a small change!