Added first version of SimpleReactor class to solver.
The SimpleReactor class is designed to provide the same functionality as
the current solver in RMG Java: a homogeneous, isothermal, isobaric
batch reactor. The temperature and pressure independence means we can
precompute the k(T) and k(T,P) values before the start of each
simulation, which will make the solver very fast, even for large kinetic
models.

Currently the class is implemented in pure Python until it is
established that it is working. Later we will almost certainly need to
port it to Cython for speed.