Added first version of a unit test for the SimpleReactor class.
Although this test doesn't assert anything, it is still useful in that
it passes as long as the solver initializes and runs successfully. It
also produces a plot of the solver output so you can visually check that
it was successful. Later we may want to add some assertions regarding
the simulation output.