Improved method of specifying parameter for DASSL.step().
We don't want to choose a time too far in the future, or DASSL won't run
correctly. Instead we increment the parameter by an order of magnitude
once we get close to its value so it's always pretty close to the
current simulation time. Note that we are still allowing the solver to
choose the time steps.

In the future we may want to do something else each time we advance the
step parameter, e.g. print the simulation status to the log file.