Lots of work toward Cythonizing of rmgsolver.base and rmgsolver.simple.
Particular focus was made on the SimpleReactor.residual() function, as
that is almost certainly going to be the bottleneck of the simulation
process. I've also turned off bounds checking for the arrays, which,
when combined with declaring of the stride of the numpy arrays and the
type of all indices, really minimizes the amount of C code generated.

Note that we must provide pydas.pxd in order for everything to build
properly, and that it must be in the root solver directory so the
Cythonized modules know to look in the global namespace for the pydas
module.

I've had to rename a few things so that they don't trample on names in
the DASSL base class.

The speed increase I'm seeing as a result of this commit alone is
tremendous. I haven't tried to quantify it yet, but I'm guessing an
order of magnitude isn't too much of an exaggeration.