Removed cantera-related code from all projects.
The idea here is to remove cantera as a dependency for RMG Py, because
it is nontrivial to install and doesn't really provide any killer
features. In particular, although it allows for several types of
reaction systems, it only uses VODE as the solver, which won't be nearly
robust enough for large kinetic models. The new rmgsolver subproject
will provide this functionality using DASSL as the solver.

We will need to implement another feature that cantera provided: writing
of Chemkin files.

a built-in solver allowing for
several types of reaction systems, it only uses VODE