Now using rmgsolver within RMG Py.
The BatchReactor class from the (soon to be deprecated) rmg.system.batch
module has been replaced with the SimpleReactor class from the
rmgsolver.simple module. The latter is actually simpler to use because
it only requires T, P, and initial mole fractions. The input file
reading has been updated accordingly.

The switch to the rmgsolver package provides a tremendous speed boost to
RMG Py, enough to bring it within a factor of two of RMG Java (at least with
pressure dependence off). This means that speed is no longer a reason not to
make the switch!