Rewrite of addReactionToUnimolecularNetworks() to avoid merging of networks.
The previous algorithm would merge networks that contained any common
explored unimolecular wells. This has been observed to generate many
very large networks that are computationally expensive to run and give a
lot of k(T,P) values that are negligible, meaning most of that time is
actually wasted.

The idea is to only merge partial networks if they contain one or more
common unimolecular wells *and* the same source configuration. This is
much closer to the original "activated species algorithm" of Matheu et
al.

This commit is the first step towards implementation of the new
algorithm. It provides a new function for sorting unimolecular reactions
into networks that distinguishes networks by source as well as explored
wells.