Rewrite of updateUnimolecularReactionNetworks() to reflect new algorithm.
The key idea here is that each partial network is responsible for
computing k(T,P) values for reactions involving its source configuration
as the reactant only. Thus we should only have 0 or 1 reactant channel,
and the only isomers should be the explored wells + the source (if
unimolecular).

The bulk of the new algorithm was moved into the PDepNetwork.update()
method for readability.