Only generate frequencies when needed for pressure dependence code.
The frequency-generating algorithm is quite approximate, and fairly
expensive, so we want to minimize the number of times we do it. However,
we were previously generating frequencies for all species in the core
and edge. Now we only generate frequencies for species that are
reactants or explored isomers (which are the only ones we need
frequencies for when minimizeDensityOfStatesCalculations is True).